Mostly a strategic refactor.
Moved setup_rules_and_policies into its own file 'rules', which is
referred to from the main script as the 'optional' $RULES_FILE

This makes it easier to separate the 'static' parts from the 'dynamic'
rules. Also makes it easier to provide more than one $RULES_FILE as
sample/example configurations, as it's becoming difficult to show all
functionality in a single, meaningful, example.

If the rules file doesn't exist, glars defaults to the same behaviour as
a factory-reset router.