Added support for logging. Now you can specify LOG=1 to enable logging to dmesg. Dropped packets (either due to rate limiting or blacklisting) will be logged. 'dmesg|grep GLARS' to view the logs .
Also fixed a pretty bad bug in the blacklisting code. Traffic coming in
from blacklisted IPs to ports that are forwarded was NOT being dropped,
because the DROP rules is needed for the FORWARD chain in addition to the
INPUT chain.

Also fixed limit_connections_to_internal_host() so it now works as
expected, and allowed comments in the blacklist file. Also bumped
the version to GLARS 2.0 - about time right? Maybe some other minor
cleanups.